Use a regex for isBcryptHash
This improves the correctness of the test. The current check allows both invalid message (invalid character set) and an invalid cost (more than 2 digits).

It should also improve the performance of the test. Splitting the hash and testing each part is likely slower than the pre-compiled regular expression, and must evaluate more conditions to fail.